fix(ci): tolerate transient chat rate-limits in e2e
Two-layer defense against Google-side chat throttling on the CI account,
which has been turning ~7 e2e tests red per nightly even though the same
queries pass on other accounts at the same wall-clock time:

1. Promote the `_skip_on_chat_rate_limit` wrapping from `tests/e2e/test_chat.py`
   into `tests/e2e/conftest.py`, installed by the `client` fixture so every
   e2e test that uses `client.chat.ask` gets the same auto-skip behavior.
   Now test_source_selection.py and test_notebooks.py also degrade rate-limit
   ChatErrors to skips instead of failures.

2. Replace `--reruns 2 --reruns-delay 30` (which sleeps per-test attempt) with
   `--reruns 1 --reruns-delay 30` plus a single `if: steps.e2e.outcome == 'failure'`
   retry step that sleeps 600s once and runs `pytest --last-failed`. The
   skip-fixture stays active during the retry, so tests still throttled at
   +10 min become skips rather than red CI.

   Initial e2e step is `continue-on-error: true` so a passing retry actually
   greens the job. The retry guards against missing lastfailed cache (pytest
   crash before first cache write) so it doesn't accidentally re-run the
   full ~30-min suite.

Applied to both nightly.yml and verify-package.yml. Verified locally with
peopleconf profile: all 5 previously-failing tests pass cleanly under the
new fixture.
